+++ tests/testthat/test-rectify.R
@@ -194,10 +194,10 @@ test_that("rectify() works on common data types", {
       lgl = c(TRUE, FALSE),
       int = c(1L, 2L),
       dbl = c(1, 2),
-      cpl = c(1i, 2i),
+      cplx = c(1i, 2i),
       date = c(as.Date("2001-01-01"), as.Date("2001-01-02")),
       chr = c("a", "b"),
-      fct = factor(c("c", "d")),
+      fctr = factor(c("c", "d")),
       list = list(1:2, letters[1:2])
     )
   y <-